yocto/meta-altera: reintroduce meta-altera
Currently 2019 no special kernel or specially patched u-boot is needed to
support the cyclone5. Current mainlines for kernel and bootloader should have
the correct settings available. A yocto layer introducing patches on those is
thus not needed.

There are certain general settings, though, which makes it a bit nasty to simply
forget the meta-altera, re-add still what is missing in terms of wic and kernel
version.

To avoid that work now, but to have a working toolchain in order to setup
manually a u-boot bootloader, the easiest way was to re-install meta-altera and
have a more specific look into it later.